requirements.txt
spotipy
plotly
numpy
matplotlib
seaborn
jupyter
pandas
python-dotenv
scikit-learn
